apps/recommender/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
numpy==2.2.3
scipy==1.15.2
python-dotenv==1.0.1
httpx==0.28.1
